airflow/pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.7",
    "apache-airflow-providers-google>=19.5.0",
    "apache-airflow-providers-postgres>=6.5.4",
    "google-cloud-storage>=3.9.0",
]
